ref(client): Breaks up CLI client into separate modules
The client.rs file was getting a tad big, so I pulled the options out
into their own file. The files are pretty much the same otherwise. So while
this looks big, it isn't
