+++ Cargo.toml
@@ -74,7 +74,7 @@ required-features = ["cli"]
 
 [[bin]]
 name = "bindle"
-path = "bin/client.rs"
+path = "bin/client/main.rs"
 required-features = ["cli"]
 
 [[bin]]
